updated environment variable

+++ PathLDM/main.py
@@ -764,7 +764,7 @@ if __name__ == "__main__":
         if opt.train:
             try:
                 # 
-                trainer.fit(model, data)
+                trainer.fit(model.half(), data)
             except Exception:
                 print(type(model))
                 print(type(data))
